requirements.txt
nidaqmx
pyserial
scipy
